feat: 북마크 목록 조회 기능 개발 (#299)

+++ api/src/main/java/com/sponus/sponusbe/domain/bookmark/service/BookmarkService.java
@@ -10,6 +10,8 @@ import com.sponus.coredomain.domain.organization.repository.OrganizationLinkRepo
 import com.sponus.coredomain.domain.organization.repository.OrganizationRepository;
 import com.sponus.sponusbe.domain.bookmark.dto.request.BookmarkToggleRequest;
 import com.sponus.sponusbe.domain.bookmark.dto.response.BookmarkToggleResponse;
+import com.sponus.sponusbe.domain.bookmark.exception.BookmarkErrorCode;
+import com.sponus.sponusbe.domain.bookmark.exception.BookmarkException;
 import com.sponus.sponusbe.domain.organization.exception.OrganizationErrorCode;
 import com.sponus.sponusbe.domain.organization.exception.OrganizationException;
 
@@ -26,16 +28,22 @@ public class BookmarkService {
 	private final OrganizationRepository organizationRepository;
 
 	public BookmarkToggleResponse bookmarkToggle(Organization organization, BookmarkToggleRequest request) {
-		return bookmarkRepository.findByOrganization(organization)
+
+		if (organization.getId().equals(request.target()))
+			throw new BookmarkException(BookmarkErrorCode.BOOKMARK_ERROR);
+
+		final Organization target = organizationRepository.findById(request.target())
+			.orElseThrow(() -> new OrganizationException(OrganizationErrorCode.ORGANIZATION_NOT_FOUND));
+		return bookmarkRepository.findByOrganizationAndTarget(organization, target)
 			.map(existingBookmark -> {
 				bookmarkRepository.delete(existingBookmark);
 				return BookmarkToggleResponse.from(existingBookmark, false);
 			})
 			.orElseGet(() -> {
-				final Organization target = organizationRepository.findById(request.target())
-					.orElseThrow(() -> new OrganizationException(OrganizationErrorCode.ORGANIZATION_NOT_FOUND));
 				final Bookmark bookmark = bookmarkRepository.save(request.toEntity(organization, target));
 				return BookmarkToggleResponse.from(bookmark, true);
 			});
 	}
+
+
 }
